Adding hybrid tag to inverter
Hybrid inverter do not provide run_state. Therefor it need to determined by
system_state.
In a first step a list of gridonly and hybrid inverters was added to register.yaml
setup.py was adapted in order to overwrite run_state by system_state if
it is a hybrid inverter.

ToDo: In future parsing of the states is nescarry.