Make pvoutput reports align with status interval wall clock time
The previous code had a problem where it would do an intitial report
on start up and then add status_interval minutes to the next report
time *after* it had finished processing.

This causes 2 problems:
 1. Time drift - the report time interval is longer than the actual
    status_interval so over the course of the day ends up being later
    and later which can lead to a gap in the pvoutput data.
 2. The report time is unpredictable, which will cause clashes with other
    systems that report consumption data each time the report time
    drifts enough to clash with the other system's upload with the
    end result that one cancels out the other.

To fix #2, the second system just needs to report the same report
timestamp 30 seconds later than the first.

The fix here aligns the report time to a rounded fraction of the
wall clock time so that it matches the wall clock time on pvoutput's
list of reports on the UI. Typically if you report every 5 minutes,
it will now send a report to pvoutput at :00, :05, :10, etc.
instead of the previous non-rounded times that drifted forward.